Major update of ASP code
Finally succeed to perform the long-awaited modifications:

- get rid of cc constant, that was useless

- reorganize code, delete legacy and useless lines

- The number of atoms in output is drastically reduced by indicating
which part of the hierarchy tree is modified, instead of
outputing all the block hierarchy at each step.
This is indicated with hierarchy_add/4 and hierarchy_remove/4 atoms.
It requires then to be handled python-side to maintain a valid hierarchy: commits are coming.

- powernode reuse is properly detected and handled, so that outside
world do not even know that an already used powernode was somehow
handled as a new one by ASP. Outside world just receive new powernodes
and poweredges.