ASP do not yield nodes of already existant powernode + simplify edge cover computation in searchers
this is a small improvement: now nodes are only generated once in ASP:
when they are pushed in a new powernode.
Existing pnodes are just mentioned, ang Graph object provides
motif.edge_covered method with the sets of nodes implied in the motif.
Therefore, MotifSearcher instances do not need to access motif/ASP model
to get infos about the motif, and can just do
their computations on sets of nodes.