fix test case: different handling when StarSearch is involved
This is due to incoming bounds computation fix, that in this
particular case, disable search for cliques.
Results is that the clique b2 b3 b4 is not searched, and the star
b1 × b2 b3 b4 is taken instead, as a biclique by fullbiclique search.
Consequence : b2 b3 b4 is second set instead of first one, leading to an
error in the fourth test run (that involve fullbiclique search).

I'm unsure about that. Is that a bug in the code, or a bug in the tests
?
Well, it's certainly not a problem for the code if the bounds for motifs
are fixed. But the test case «motif overlapping» really deserve its
name, now.