the first motif to reach the best score is compressed, instead of the last
this is done by considering the incremented score to beat instead of the
score to beat raw. This way, if a previous motif of score 7 has been
found, i have to report a motif of score 8 or better to be accepted, not
7.
This changes the order of compression for some motifs, because now the
first best motif to be found is also the one to be compressed.

This broke a test, where that order was important.
See 29ac96da6996b9df2f18ac74d0dc107069fa7c3e